+++ crates/payload/util/src/lib.rs
@@ -11,5 +11,5 @@
 mod traits;
 mod transaction;
 
-pub use traits::{BestPayloadTransactions, NoopPayloadTransactions, PayloadTransactions};
+pub use traits::{NoopPayloadTransactions, PayloadTransactions};
 pub use transaction::{PayloadTransactionsChain, PayloadTransactionsFixed};
